requirements.txt
flask
supabase
python-dotenv
werkzeug
openpyxl
reportlab
gunicorn
weasyprint